Improvements to custom weapons, character combat, small fixes
* Improved custom weapons now includes muzzle, smoke and brass
* Improved response for character combat, damage decals for melee
* Debug visual for firespot of character weapons
* Dropped weapons are now true objects with running weapon scripts
* Capped lightspot radius
* Standalone now retains HUD when stock HUD system is used
* Fixed issue of clip ammo being wiped out
* No longer deletes collection files from root when saving to writables
* Health panel and values now hide when player is invulnerable or briefly immune

+++ GameGuru Core/GameGuru/Source/M-RPG.cpp
@@ -346,7 +346,9 @@ bool save_rpg_system_items(char* name, bool bIncludeELEFile)
 	strcpy(collectionfilename, "projectbank\\");
 	strcat(collectionfilename, name);
 	strcat(collectionfilename, "\\collection - items.tsv");
-	DeleteFileA(collectionfilename);
+	//DeleteFileA(collectionfilename);
+	GG_GetRealPath(collectionfilename, 1);
+	if (FileExist(collectionfilename) == 1) DeleteFileA(collectionfilename);
 	FILE* collectionFile = GG_fopen(collectionfilename, "w");
 	if (collectionFile)
 	{
@@ -465,7 +467,9 @@ bool save_rpg_system_quests(char* name)
 	strcpy(collectionfilename, "projectbank\\");
 	strcat(collectionfilename, name);
 	strcat(collectionfilename, "\\collection - quests.tsv");
-	DeleteFileA(collectionfilename);
+	//DeleteFileA(collectionfilename);
+	GG_GetRealPath(collectionfilename, 1);
+	if (FileExist(collectionfilename) == 1) DeleteFileA(collectionfilename);
 	FILE* collectionFile = GG_fopen(collectionfilename, "w");
 	if (collectionFile)
 	{
@@ -878,7 +882,10 @@ bool refresh_collection_from_entities(void)
 						break;
 					}
 				}
-				if ( bFoundAndAssignedE == false ) bNeedMoreEntitiesLoading = true;
+				if (bFoundAndAssignedE == false)
+				{
+					bNeedMoreEntitiesLoading = true;
+				}
 			}
 		}
 	}
